# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HandoffRail",
    "description": "Session-continuity middleware for multi-agent AI workflows",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/keys": {
      "get": {
        "tags": [
          "keys"
        ],
        "summary": "List Api Keys",
        "description": "List all API keys for the current tenant. The actual key values are not shown.",
        "operationId": "list_api_keys_api_v1_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponse"
                  },
                  "type": "array",
                  "title": "Response List Api Keys Api V1 Keys Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key. Requires an existing valid API key.",
        "operationId": "create_api_key_api_v1_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/keys/{key_id}": {
      "delete": {
        "tags": [
          "keys"
        ],
        "summary": "Revoke Api Key",
        "description": "Revoke (soft-delete) an API key. The key is marked as revoked, not removed.\n\nOnly keys belonging to the same tenant as the current key can be revoked,\nunless the current key is an admin key.",
        "operationId": "revoke_api_key_api_v1_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/keys/{key_id}/rotate": {
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "Rotate Api Key",
        "description": "Rotate an API key \u2014 creates a new key and revokes the old one.\n\nThe new key inherits the same name (with ' (rotated)' suffix), tenant_id,\nand tier as the original. The old key is immediately revoked.\n\nOnly keys belonging to the same tenant can be rotated, unless the current\nkey is an admin key.\n\nReturns the new key with the plain key value (shown only once).",
        "operationId": "rotate_api_key_api_v1_keys__key_id__rotate_post",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tenants/{tenant_id}/keys": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "List Tenant Keys",
        "description": "List all API keys for a specific tenant. Requires admin API key.",
        "operationId": "list_tenant_keys_api_v1_tenants__tenant_id__keys_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponse"
                  },
                  "title": "Response List Tenant Keys Api V1 Tenants  Tenant Id  Keys Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ApiKeyCreate",
        "description": "Request body for creating a new API key."
      },
      "ApiKeyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin",
            "default": false
          },
          "revoked": {
            "type": "boolean",
            "title": "Revoked"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "key_prefix",
          "tenant_id",
          "revoked",
          "created_at"
        ],
        "title": "ApiKeyResponse",
        "description": "Response for an API key (includes the plain key only on creation)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}